pyproject.toml
[project]
name = "shadowstep-demo"
version = "0.0.0"
requires-python = ">=3.9"

dependencies = ["appium-python-client-shadowstep==0.34.165"]
